refactor: move configmap functions into utils package
